Remove unnecessary buffer copy in the packet decoding and disassembler packet data directly from the ByteReadChannel.

+++ script/src/main/kotlin/script/packet/disassembler/WindowStatusDisassembler.kts
@@ -11,7 +11,7 @@ import xlitekt.shared.buffer.readUShort
  */
 onPacketDisassembler(opcode = 48, size = 5) {
     WindowStatusPacket(
-        displayMode = readByte(),
+        displayMode = readByte().toInt(),
         width = readUShort(),
         height = readUShort()
     )
